Removed legacy menu
