+++ Fedoraware/TeamFortress2/TeamFortress2.vcxproj.filters
@@ -11,7 +11,6 @@
     <ClCompile Include="TeamFortress2\Features\ESP\ESP.cpp" />
     <ClCompile Include="TeamFortress2\Features\Glow\Glow.cpp" />
     <ClCompile Include="TeamFortress2\Features\Menu\InputHelper\InputHelper.cpp" />
-    <ClCompile Include="TeamFortress2\Features\Menu\Menu.cpp" />
     <ClCompile Include="TeamFortress2\Features\Misc\Misc.cpp" />
     <ClCompile Include="TeamFortress2\Features\Chams\Chams.cpp" />
     <ClCompile Include="TeamFortress2\Features\Prediction\Prediction.cpp" />
@@ -66,16 +65,16 @@
     <ClCompile Include="TeamFortress2\Features\Keybinds\Keybinds.cpp" />
     <ClCompile Include="TeamFortress2\Features\Console\Console.cpp" />
     <ClCompile Include="TeamFortress2\Features\PlayerArrows\PlayerArrows.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_demo.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_draw.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_impl_dx9.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_impl_win32.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_tables.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_widgets.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\What.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_demo.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_draw.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_impl_dx9.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_impl_win32.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_tables.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_widgets.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\Menu.cpp" />
     <ClCompile Include="TeamFortress2\Hooks\MenuHook\MenuHook.cpp" />
-    <ClCompile Include="TeamFortress2\Features\What\ImGui\imgui_stdlib.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Menu\ImGui\imgui_stdlib.cpp" />
     <ClCompile Include="TeamFortress2\SDK\Interfaces\NetChannel\NetChannel.cpp" />
     <ClCompile Include="TeamFortress2\Utils\Events\Events.cpp" />
     <ClCompile Include="TeamFortress2\Features\Crits\Crits.cpp" />
@@ -106,7 +105,6 @@
     <ClInclude Include="TeamFortress2\Features\ESP\ESP.h" />
     <ClInclude Include="TeamFortress2\Features\Glow\Glow.h" />
     <ClInclude Include="TeamFortress2\Features\Menu\InputHelper\InputHelper.h" />
-    <ClInclude Include="TeamFortress2\Features\Menu\Menu.h" />
     <ClInclude Include="TeamFortress2\Features\Misc\Misc.h" />
     <ClInclude Include="TeamFortress2\Features\Chams\Chams.h" />
     <ClInclude Include="TeamFortress2\Features\Prediction\Prediction.h" />
@@ -467,15 +465,15 @@
     <ClInclude Include="TeamFortress2\Features\Keybinds\Keybinds.h" />
     <ClInclude Include="TeamFortress2\Features\Console\Console.h" />
     <ClInclude Include="TeamFortress2\Features\PlayerArrows\PlayerArrows.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imconfig.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imgui.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imgui_impl_dx9.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imgui_impl_win32.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imgui_internal.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imstb_rectpack.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imstb_textedit.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imstb_truetype.h" />
-    <ClInclude Include="TeamFortress2\Features\What\What.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imconfig.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imgui.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imgui_impl_dx9.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imgui_impl_win32.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imgui_internal.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imstb_rectpack.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imstb_textedit.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imstb_truetype.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\Menu.h" />
     <ClInclude Include="TeamFortress2\Hooks\MenuHook\MenuHook.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\d3dx9.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\DirectX.h" />
@@ -489,7 +487,7 @@
     <ClInclude Include="TeamFortress2\SDK\DirectX\d3dx9effect.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\d3dx9math.h" />
     <ClInclude Include="TeamFortress2\Utils\MinHook\hook.h" />
-    <ClInclude Include="TeamFortress2\Features\What\ImGui\imgui_stdlib.h" />
+    <ClInclude Include="TeamFortress2\Features\Menu\ImGui\imgui_stdlib.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\Include\audiodefs.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\Include\comdecl.h" />
     <ClInclude Include="TeamFortress2\SDK\DirectX\Include\D2D1.h" />
