+++ Fedoraware/TeamFortress2/TeamFortress2/Features/Menu/ConfigManager/ConfigManager.cpp
@@ -525,7 +525,6 @@ void CConfigManager::Save(const wchar_t *name)
 			SAVE_VAR(Vars::Misc::BeCat);
 			SAVE_VAR(Vars::Misc::VoteRevealer);
 			SAVE_VAR(Vars::Misc::VotesInChat);
-			SAVE_VAR(Vars::Menu::LegacyMenu);
 			// CL_Move
 			{
 				SAVE_VAR(Vars::Misc::CL_Move::Enabled);//Enabled
@@ -1032,7 +1031,6 @@ void CConfigManager::Load(const wchar_t *name)
 			LOAD_VAR(Vars::Visuals::MaxDist);
 			LOAD_VAR(Vars::Visuals::MinDist);
 			LOAD_VAR(Vars::Visuals::AimPosSquare);
-			LOAD_VAR(Vars::Menu::LegacyMenu);
 			LOAD_VAR(Vars::Visuals::despawnTime);
 			LOAD_VAR(Vars::Visuals::damageLogger);
 			LOAD_VAR(Vars::Glow::Main::Wireframe);
