+++ Fedoraware/TeamFortress2/TeamFortress2/Features/Menu/Menu.h
@@ -1,41 +1,29 @@
 #pragma once
-
-#include "../Vars.h"
+#include "ImGui/imgui_impl_win32.h"
+#include "ImGui/imgui_impl_dx9.h"
+#include "../../SDK/SDK.h"
 
 class CMenu
 {
 private:
-	Rect_t m_LastWidget = {};
-	Rect_t m_LastGroupBox = {};
-
-	void Separator();
-	bool CheckBox(CVar<bool>& Var, const wchar_t* const szToolTip);
-
-	bool ComboBox(CVar<int>& Var, const std::vector<CVar<int>>& List);
-	bool InputFloat(CVar<float>& Var, float Min, float Max, float Step = 1.0f, const wchar_t* Fmt = L"%f");
-	bool InputInt(CVar<int>& Var, int Min, int Max, int Step = 1);
-	bool InputColor(Color_t& Var, const wchar_t* Label);
-	bool InputString(const wchar_t* Label, std::wstring& output);
-	bool InputConstChar(const wchar_t* Label, std::string& output);
-	bool InputCString(const wchar_t* Label, std::string& output);
-	bool InputKey(CVar<int>& output, bool bAllowNone = true);
-	void GroupBoxStart();
-	void GroupBoxEnd(const wchar_t* Label, int Width);
-	void DrawTooltip();
-
-private:
-	bool m_bReopened = false;
-	std::wstring m_szCurTip = L"";
+	ImFont* Normal;
+	ImFont* DT;
+	ImDrawList* drawList;
 
 public:
-	bool Button(const wchar_t* Label, bool Active = false, int WidthOverride = 0, int HeightOverride = 0);
+	bool config = false;
+	bool menuOpen = false;
+	bool menuUnload = false;
+	void TextCenter(std::string text);
+	void Render(IDirect3DDevice9* pDevice);
+
+public:
+	bool m_bReopened = false;
 	float m_flFadeAlpha = 0.0f;
 	float m_flFadeElapsed = 0.0f;
 	float m_flFadeDuration = 0.1f;
 	float flTimeOnChange = 0.0f;
-public:
 	bool m_bOpen = false;
-	bool m_bTyping = false;
 	void Run();
 };
