+++ Fedoraware/TeamFortress2/TeamFortress2/Hooks/SurfaceHook/SurfaceHook.cpp
@@ -1,6 +1,5 @@
 #include "SurfaceHook.h"
 #include "../../Features/Menu/Menu.h"
-#include "../../Features/What/What.h"
 
 void __stdcall SurfaceHook::OnScreenSizeChanged::Hook(int OldWidht, int OldHeight)
 {
@@ -12,6 +11,5 @@ void __stdcall SurfaceHook::OnScreenSizeChanged::Hook(int OldWidht, int OldHeigh
 
 void __stdcall SurfaceHook::LockCursor::Hook()
 {
-	g_Menu.m_bOpen ? g_Interfaces.Surface->UnlockCursor() : Table.Original<fn>(index)(g_Interfaces.Surface);
-	g_What.menuOpen ? g_Interfaces.Surface->UnlockCursor() : Table.Original<fn>(index)(g_Interfaces.Surface);
+	g_Menu.menuOpen ? g_Interfaces.Surface->UnlockCursor() : Table.Original<fn>(index)(g_Interfaces.Surface);
 }
